requirements.txt
# Core inferencia
transformers>=4.56,<5
huggingface-hub<1.0
accelerate>=0.27.0
sentencepiece>=0.2.0
torch>=2.5,<2.11
bitsandbytes>=0.49.2
llama-cpp-python>=0.3.20
vllm>=0.19.0; platform_system == "Linux"
llguidance>=0.7.30; platform_system == "Linux"
xgrammar>=0.1.25; platform_system == "Linux"
ninja
# Medición energía / hardware
codecarbon>=2.3.1
nvidia-ml-py>=12.560.30

# Evaluación / métricas
datasets>=2.18.0
human-eval>=1.0.3
pandas>=2.2.3,<2.4
numpy<=2.2
lm_eval
langdetect
immutabledict
evaluate
rouge_score
# Paquete local
-e .
